[refactor] Handle L2 detach in VM cascade
