+++ compute/src/main/java/org/zstack/compute/vm/VmCascadeExtension.java
@@ -24,6 +24,9 @@ import org.zstack.header.identity.AccountResourceRefVO;
 import org.zstack.header.identity.AccountResourceRefVO_;
 import org.zstack.header.identity.AccountVO;
 import org.zstack.header.message.MessageReply;
+import org.zstack.header.network.l2.L2NetworkConstant;
+import org.zstack.header.network.l2.L2NetworkDetachStruct;
+import org.zstack.header.network.l2.L2NetworkVO;
 import org.zstack.header.network.l3.IpRangeInventory;
 import org.zstack.header.network.l3.IpRangeVO;
 import org.zstack.header.network.l3.L3NetworkInventory;
@@ -42,6 +45,7 @@ import org.zstack.utils.function.Function;
 import org.zstack.utils.logging.CLogger;
 
 import javax.persistence.Query;
+import javax.persistence.Tuple;
 import javax.persistence.TypedQuery;
 import java.util.*;
 import java.util.concurrent.Callable;
@@ -120,9 +124,60 @@ public class VmCascadeExtension extends AbstractAsyncCascadeExtension {
             handleDeletionCleanup(action, completion);
         } else if (action.isActionCode(PrimaryStorageConstant.PRIMARY_STORAGE_DETACH_CODE)) {
             handlePrimaryStorageDetach(action, completion);
+        } else if (action.isActionCode(L2NetworkConstant.DETACH_L2NETWORK_CODE)) {
+            handleL2NetworkDetach(action, completion);
         } else {
             completion.success();
         }
+    }
+
+    @Transactional(readOnly = true)
+    private List<DetachNicFromVmMsg> getVmNicDetachMsgs(List<L2NetworkDetachStruct> structs) {
+        List<DetachNicFromVmMsg> dmsgs  = new ArrayList<>();
+
+        for (L2NetworkDetachStruct s : structs) {
+            String sql = "select vm.uuid, nic.uuid from VmInstanceVO vm, VmNicVO nic, L3NetworkVO l3"
+                    + " where vm.clusterUuid = :clusterUuid"
+                    + " and l3.l2NetworkUuid = :l2NetworkUuid"
+                    + " and nic.l3NetworkUuid = l3.uuid "
+                    + " and nic.vmInstanceUuid = vm.uuid";
+            TypedQuery<Tuple> q = dbf.getEntityManager().createQuery(sql, Tuple.class);
+            q.setParameter("clusterUuid", s.getClusterUuid());
+            q.setParameter("l2NetworkUuid", s.getL2NetworkUuid());
+            dmsgs.addAll(q.getResultList().stream().map((t) -> {
+                DetachNicFromVmMsg msg = new DetachNicFromVmMsg();
+                msg.setVmInstanceUuid(t.get(0, String.class));
+                msg.setVmNicUuid(t.get(1, String.class));
+                bus.makeTargetServiceIdByResourceUuid(msg, VmInstanceConstant.SERVICE_ID, msg.getVmInstanceUuid());
+                return msg;
+            }).collect(Collectors.toList()));
+        }
+
+        return dmsgs;
+    }
+
+    private void handleL2NetworkDetach(CascadeAction action, final Completion completion) {
+        List<L2NetworkDetachStruct> structs = action.getParentIssuerContext();
+        final List<DetachNicFromVmMsg> dmsgs = getVmNicDetachMsgs(structs);
+        if (dmsgs.isEmpty()) {
+            completion.success();
+            return;
+        }
+
+        bus.send(dmsgs, 20, new CloudBusListCallBack(completion) {
+            @Override
+            public void run(List<MessageReply> replies) {
+                for (MessageReply r : replies) {
+                    if (!r.isSuccess()) {
+                        DetachNicFromVmMsg msg = dmsgs.get(replies.indexOf(r));
+                        logger.warn(String.format("failed to stop vm[uuid:%s] for l2Network detached, %s." +
+                                " However, detaching will go on", msg.getVmInstanceUuid(), r.getError()));
+                    }
+                }
+
+                completion.success();
+            }
+        });
     }
 
     @Transactional(readOnly = true)
@@ -385,6 +440,7 @@ public class VmCascadeExtension extends AbstractAsyncCascadeExtension {
                 L3NetworkVO.class.getSimpleName(),
                 IpRangeVO.class.getSimpleName(),
                 PrimaryStorageVO.class.getSimpleName(),
+                L2NetworkVO.class.getSimpleName(),
                 InstanceOfferingVO.class.getSimpleName(),
                 AccountVO.class.getSimpleName());
     }
